+++ frontend_mobile/AudioScholar/app/src/main/java/edu/cit/audioscholar/data/remote/service/ApiService.kt
@@ -12,7 +12,6 @@ import okhttp3.RequestBody
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.Header
 import retrofit2.http.Multipart
 import retrofit2.http.POST
 import retrofit2.http.Part
