+++ frontend_mobile/AudioScholar/app/src/main/java/edu/cit/audioscholar/domain/repository/RemoteAudioRepository.kt
@@ -1,0 +1,22 @@
+package edu.cit.audioscholar.domain.repository
+
+import android.net.Uri
+import edu.cit.audioscholar.data.remote.dto.AudioMetadataDto
+import kotlinx.coroutines.flow.Flow
+
+sealed class UploadResult {
+    data object Success : UploadResult()
+    data class Error(val message: String) : UploadResult()
+    data class Progress(val percentage: Int) : UploadResult()
+    data object Loading : UploadResult()
+}
+
+interface RemoteAudioRepository {
+    fun uploadAudioFile(
+        fileUri: Uri,
+        title: String?,
+        description: String?
+    ): Flow<UploadResult>
+
+    fun getCloudRecordings(): Flow<Result<List<AudioMetadataDto>>>
+}
